use &str not String, move gmsh io tests to src/io/gmsh.rs
